test: Move test_help to pytest
